requirements.txt
apturl==0.5.2
asgiref==3.4.1
blinker==1.4
catfish==1.4.13
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.2.6
django-cors-headers==3.7.0
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
entrypoints==0.3
flake8==3.9.2
gunicorn==20.1.0
httplib2==0.14.0
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lightdm-gtk-greeter-settings==1.2.2
Markdown==3.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
menulibre==2.2.1
mugshot==0.4.2
netifaces==0.10.4
oauthlib==3.1.1
olefile==0.46
onboard==1.4.1
pexpect==4.6.0
Pillow==8.3.1
psutil==5.5.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycairo==1.16.2
pycodestyle==2.7.0
pycparser==2.20
pycups==1.9.73
pydocstyle==6.1.1
pyflakes==2.3.1
PyGObject==3.36.0
PyJWT==2.1.0
pymacaroons==0.13.0
PyNaCl==1.3.0
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.3.1
reportlab==3.5.34
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
sgt-launcher==0.2.5
simplejson==3.16.0
six==1.16.0
snowballstemmer==2.1.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
systemd-python==234
ubuntu-advantage-tools==27.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.26.6
wadllib==1.3.3
whitenoise==5.3.0
xcffib==0.8.1
xkit==0.0.0
